Create postOneResultError.ts
